+++ src/core/dag/ConverterNode.h
@@ -35,14 +35,14 @@
  class ConverterNode : public DeterministicNode {
 
     public:
-                                ConverterNode(const std::string& type, int dim);                            //!< Constructor of pristine node
-                                ConverterNode(DAGNode* origNode, const std::string& type, int dim);         //!< Basic constructor
+                                ConverterNode(const std::string& type, size_t dim);                         //!< Constructor of pristine node
+                                ConverterNode(DAGNode* origNode, const std::string& type, size_t dim);      //!< Basic constructor
         virtual                ~ConverterNode(void);                                                        //!< Virtual destructor
 
         // Basic utility functions
         ConverterNode*          clone(void) const;                                                          //!< Clone the converter node
         const VectorString&     getDAGClass(void) const;                                                    //!< Get DAG node class vector
-        int                     getDim(void) const { return valueDim; }                                     //!< Get dim of converted value (0 for scalar, 1 for vector, etc)
+        size_t                  getDim(void) const { return valueDim; }                                     //!< Get dim of converted value (0 for scalar, 1 for vector, etc)
         void                    printStruct(std::ostream& o) const;                                         //!< Print struct for user
         std::string             richInfo(void) const;                                                       //!< Complete info about object
 
@@ -55,7 +55,7 @@
         void                    update(void);                                                               //!< Update value and storedValue
 
         // Member variables
-        int                     valueDim;                                                                   //!< Dimensions of converted value
+        size_t                  valueDim;                                                                   //!< Dimensions of converted value
 };
 
 #endif
