+++ src/core/datatypes/primitive/Natural.h
@@ -46,12 +46,12 @@ class Natural : public Integer {
                                 operator unsigned int(void) const { return value; }             //!< Type conversion to unsigned int
 
         // Basic utility functions
-        Natural*                clone(void) const;                                              //!< Clone object
-        RbObject*               convertTo(const std::string& type, int dim) const;              //!< Convert to type and dim
-        const VectorString&     getClass(void) const;                                           //!< Get class vector
-        bool                    isConvertibleTo(const std::string& type, int dim, bool once) const;     //!< Is convertible to type and dim?
-        void                    printValue(std::ostream& o) const;                              //!< Print value (for user)
-        std::string             richInfo(void) const;                                           //!< Complete info about object
+        Natural*                clone(void) const;                                                      //!< Clone object
+        RbObject*               convertTo(const std::string& type, size_t dim) const;                   //!< Convert to type and dim
+        const VectorString&     getClass(void) const;                                                   //!< Get class vector
+        bool                    isConvertibleTo(const std::string& type, size_t dim, bool once) const;  //!< Is convertible to type and dim?
+        void                    printValue(std::ostream& o) const;                                      //!< Print value (for user)
+        std::string             richInfo(void) const;                                                   //!< Complete info about object
 
         // Getters and setters
         unsigned int            getUnsignedValue(void) const { return value; }                  //!< Get value (unsigned int)
